kdt2_TIL/week18/day2/lecture_drf/requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.0.5
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.18
django-appconf==1.0.5
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-imagekit==4.1.0
djangorestframework==3.14.0
drf-yasg==1.21.5
executing==0.8.3
Faker==13.3.5
idna==3.4
inflection==0.5.1
ipython==8.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.3
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.4
stack-data==0.2.0
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.5.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.5
